Add trino group for CCX access requests (#315)
This adds two new groups in Trino: ccx-datalake-access and
ccx-sensitive-datalake-access, with access to the CCX and CCX_Sensitive
schemas, respectively. These are the same groups that we have defined in
Rover and that we use to manage access to the CCX Superset roles. This
should simplify the onboarding process for granting access to CCX data.