Update permissions for trino catalogs
This change simplifies our permissions around Trino catalogs. It makes
it so that any logged in user has read-only access to the list of
catalogs.

This will not wholesale grant all users access to all data as we have
restrictions further down the file on schema and table level access.