Generate Trino group mapping
DataHub team is working on getting LDAP groups support to Trino.
As a preparation step I unified the CCX related Rover groups
and group mapping in Trino config.

With this PR I'm proposing to use the roevr groups to drive access to
Trino as well as to Superset. To make this easier before Trino gets
LDAP support I prepared a Makefile that updates the group mapping
from the Rover group.

```
$ make help
update-ccx-groups              Update ccx trino access groups from LDAP
$ make update-ccx-groups
```

The procedure to onboard new user to Trino should be to update the rover
group, call the `make update-ccx-groups` and submit a PR.

The important thing to watch is not to allow manual addition to the group
mapping without addition to the Rover group.