Use env vars for differing settings in trino config
We need to tweak a few specific values in the trino-config secret
between dev, stage, and prod.

Previously we were doing this with patches in each overlay, but this was
very complex due to the number of places we modify this secret. We need
to start making changes to this secret that are encrypted. I think this
change is necessary for us to be able to do so.