Allow the trino-admin user to impersonate any user in trino
In order for the LDAP group provider in Trino to work, Superset needs to
connect to Trino using a valid user that exists in LDAP. We're therefore
switch to using the trino-admin user for this connection instead of the
currently used superset user. The trino-admin user then needs
permissions to impersonate any logged in user.