Grant impersonation permissions to new superset user
We need to configure Superset to use a valid LDAP user for connecting to
Trino. To prevent unauthorized data access, this user should have no
permissions in trino other than user impersonation, so we can't use the
existing trino-admin LDAP user. This change grants impersonation
permissions to the new data-hub-superset-trino user that we got created
for this purpose.